+++ CHANGELOG.md
@@ -4,6 +4,6 @@
 
 Breaking changes:
 
-- `CozyTelemetry.Metrics` and `CozyTelemetry.Measurements` are merged as `CozyTelemetry.Spec`.
+- `NeedleTelemetry.Metrics` and `NeedleTelemetry.Measurements` are merged as `NeedleTelemetry.Spec`.
 - `:metrics` option and `:measurements` option are merged as `:specs`.
 - `:optional_metrics` option and `:optional_measurements` option are merged as `:optional_specs`.